Revert "commands to arguments"

+++ s3_backup/manage.py
@@ -1,6 +1,8 @@
 import os
 import shutil
 
+import click
+
 from .installation import Installation
 
 
@@ -18,6 +20,12 @@ class ManageProfiles(Installation):
             )
 
     def do_add(self, name: str):
+        if name in self.installed_profiles:
+            raise click.BadArgumentUsage(
+                "A profile with this name has already been created. Please choose a "
+                "different name."
+            )
+
         path = self.get_profile_path(name)
         with open(self._template_path, "r") as t, open(path, "w") as n:
             n.write(t.read())
@@ -26,9 +34,14 @@ class ManageProfiles(Installation):
             os.makedirs(self.get_log_path(name))
 
     def do_edit(self, name: str):
+        if name not in self.installed_profiles:
+            raise click.BadArgumentUsage("A profile with this name does not exist.")
         os.system(f"vim {self.get_profile_path(name)}")
 
     def do_remove(self, name: str):
+        if name not in self.installed_profiles:
+            raise click.BadArgumentUsage("A profile with this name does not exist.")
         os.system(f"rm {self.get_profile_path(name)}")
+
         if os.path.isdir(self.get_log_path(name)):
             shutil.rmtree(self.get_log_path(name))
